+++ htmxology/src/lib.rs
@@ -24,7 +24,7 @@ pub use caching::{
     ControllerExt as CachingControllerExt,
 };
 pub use controller::{Controller, HasSubcontroller, SubcontrollerExt};
-pub use route::{Route, RouteExt, decode_path_argument, replace_request_path};
+pub use route::{ParseError, Route, RouteExt, decode_path_argument, replace_request_path};
 #[cfg(feature = "auto-reload")]
 pub use server::auto_reload::get_or_bind_tcp_listener;
 pub use server::{